+++ src/data/league/simulator/country_Names_Abrv.ts
@@ -1,0 +1,44 @@
+//Used to see country abbreviations and get full names from abbreviations
+export const NHL_COUNTRY_ABBREVIATIONS = [
+  "CAN",
+  "USA",
+  "SWE",
+  "FIN",
+  "RUS",
+  "CZE",
+  "CHE",
+  "SVK",
+  "DEU",
+  "NOR",
+  "LVA",
+  "BLR",
+  "AUS",
+  "GBR",
+  "UZB",
+  "AUT",
+  "FRA",
+  "DNK",
+  "SVN",
+];
+
+export const NHL_COUNTRY_NAMES: { [abbr: string]: string } = {
+  CAN: "Canada",
+  USA: "United States",
+  SWE: "Sweden",
+  FIN: "Finland",
+  RUS: "Russia",
+  CZE: "Czech Republic",
+  CHE: "Switzerland",
+  SVK: "Slovakia",
+  DEU: "Germany",
+  NOR: "Norway",
+  LVA: "Latvia",
+  BLR: "Belarus",
+  AUS: "Australia",
+  GBR: "United Kingdom",
+  UZB: "Uzbekistan",
+  AUT: "Austria",
+  FRA: "France",
+  DNK: "Denmark",
+  SVN: "Slovenia",
+};
